# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for xinde.",
        "title": "xinde API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/account/approval/list": {
            "get": {
                "description": "根据前端传来的字段，返回对应的待审批列表/已同意申请/已拒绝申请的用户列表，根据创建时间降序排列",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "管理员查看用户审批列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页数，可选，默认为1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "一页的内容数量，可选，默认为设置的默认值",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/xinde_internal_dto_account.ApprovalListResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/xinde_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/xinde_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/account/list": {
            "get": {
                "description": "返回已被通过注册申请的用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "管理员查看用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页数，可选，默认为1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "一页的内容数量，可选，默认为设置的默认值",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/xinde_internal_dto_account.ListResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/xinde_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/xinde_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/company/list": {
            "get": {
                "description": "返回所有的公司信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "管理员查看公司列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页数，可选，默认为1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "一页的内容数量，可选，默认为设置的默认值",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/xinde_internal_dto_company.ListResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/xinde_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "token错误",
                        "schema": {
                            "$ref": "#/definitions/xinde_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "没有管理员权限",
                        "schema": {
                            "$ref": "#/definitions/xinde_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/xinde_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "xinde_internal_dto_account.ApprovalListData": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string",
                    "example": "宁波鲍斯产业链服务有限公司"
                },
                "created_at": {
                    "type": "string",
                    "example": "2020-09-08 09:08:09"
                },
                "email": {
                    "type": "string",
                    "example": "13800138000@qq.com"
                },
                "id": {
                    "type": "integer",
                    "example": 2
                },
                "name": {
                    "type": "string",
                    "example": "张三，真实名称"
                },
                "phone": {
                    "type": "string",
                    "example": "13800138000"
                },
                "username": {
                    "type": "string",
                    "example": "张三，账号名称"
                },
                "why": {
                    "type": "string",
                    "example": "同意该用户的申请；拒绝该用户的申请"
                }
            }
        },
        "xinde_internal_dto_account.ApprovalListPageData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/xinde_internal_dto_account.ApprovalListData"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "example": 20
                },
                "pages": {
                    "type": "integer",
                    "example": 7
                },
                "total": {
                    "type": "integer",
                    "example": 137
                }
            }
        },
        "xinde_internal_dto_account.ApprovalListResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/xinde_internal_dto_account.ApprovalListPageData"
                },
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "xinde_internal_dto_account.ListData": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string",
                    "example": "宁波鲍斯产业链服务有限公司"
                },
                "created_at": {
                    "type": "string",
                    "example": "2020-09-08 09:08:09"
                },
                "email": {
                    "type": "string",
                    "example": "13800138000@qq.com"
                },
                "id": {
                    "type": "integer",
                    "example": 2
                },
                "name": {
                    "type": "string",
                    "example": "张三，真实名称"
                },
                "phone": {
                    "type": "string",
                    "example": "13800138000"
                },
                "price_level": {
                    "type": "string",
                    "example": "price_1"
                },
                "recent_search_at": {
                    "type": "string",
                    "example": "2020-09-08 09:08:09"
                },
                "remark": {
                    "type": "string",
                    "example": "备注"
                },
                "role": {
                    "type": "string",
                    "example": "普通用户"
                },
                "search_device": {
                    "type": "string",
                    "example": "车削刀杆"
                },
                "username": {
                    "type": "string",
                    "example": "张三，账号名称"
                }
            }
        },
        "xinde_internal_dto_account.ListPageData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/xinde_internal_dto_account.ListData"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "example": 20
                },
                "pages": {
                    "type": "integer",
                    "example": 7
                },
                "total": {
                    "type": "integer",
                    "example": 137
                }
            }
        },
        "xinde_internal_dto_account.ListResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/xinde_internal_dto_account.ListPageData"
                },
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "xinde_internal_dto_company.ListData": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "浙江省宁波市奉化区江口街道聚潮路55号"
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-09-09 09:09:09"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "宁波鲍斯产业链有限公司"
                },
                "price_level": {
                    "type": "string",
                    "example": "折扣等级1 折扣等级2 折扣等级3 折扣等级4"
                }
            }
        },
        "xinde_internal_dto_company.ListPageData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/xinde_internal_dto_company.ListData"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "example": 20
                },
                "pages": {
                    "type": "integer",
                    "example": 7
                },
                "total": {
                    "type": "integer",
                    "example": 137
                }
            }
        },
        "xinde_internal_dto_company.ListResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/xinde_internal_dto_company.ListPageData"
                },
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "xinde_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据，空时不返回该字段"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                }
            }
        }
    }
}